% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-class.R
\name{block_bbox}
\alias{block_bbox}
\title{Get block bounding box}
\usage{
block_bbox(x, ...)
}
\arguments{
\item{x}{A \code{block} object.}

\item{i}{Block column index (0-based).}

\item{j}{Block row index (0-based).}
}
\value{
Numeric vector c(xmin, ymin, xmax, ymax).
}
\description{
Get block bounding box
}
