% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalblock-class.R
\name{block_dim}
\alias{block_dim}
\title{Get block dimensions}
\usage{
block_dim(x, ...)
}
\arguments{
\item{x}{A \code{gdalblock} object.}

\item{i}{Block column index (0-based).}

\item{j}{Block row index (0-based).}
}
\value{
Integer vector c(ncol, nrow).
}
\description{
Get block dimensions
}
