% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-class.R
\name{block_index}
\alias{block_index}
\title{Get block pixel/line index}
\usage{
block_index(x, ...)
}
\arguments{
\item{x}{A \code{block} object.}

\item{i}{Block column index (0-based).}

\item{j}{Block row index (0-based).}
}
\value{
Integer vector c(xoff, yoff, xsize, ysize).
}
\description{
Get block pixel/line index
}
