% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-class.R
\name{blocks_in_bbox}
\alias{blocks_in_bbox}
\title{Find blocks intersecting a bounding box}
\usage{
blocks_in_bbox(x, ...)
}
\arguments{
\item{x}{A \code{block} object.}

\item{...}{Ignored.}

\item{bbox}{Numeric vector c(xmin, ymin, xmax, ymax).}

\item{data}{If TRUE, return full block_data(). If FALSE (default), return wk::rct.}
}
\value{
A \code{wk::rct} vector or data.frame of intersecting blocks.
}
\description{
Find blocks intersecting a bounding box
}
\examples{
\dontrun{
b <- block("/path/to/raster.tif")
blocks_in_bbox(b, c(100, 200, 500, 600))
blocks_in_bbox(b, c(100, 200, 500, 600), data = TRUE)
}
}
