% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalblock-package.R
\docType{package}
\name{gdalblock-package}
\alias{gdalblock}
\alias{gdalblock-package}
\title{gdalblock: Block-Based Access to GDAL Raster Datasets}
\description{
Provides an S7 class for block-based (tiled) access to raster datasets via GDAL. Wraps 'gdalraster' to expose block structure and efficient block-level reading.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/gdalblock}
  \item Report bugs at \url{https://github.com/hypertidy/gdalblock/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
