% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-class.R
\name{blocks_in_ext}
\alias{blocks_in_ext}
\title{Find blocks intersecting an extent}
\usage{
blocks_in_ext(x, ...)
}
\arguments{
\item{x}{A \code{block} object.}

\item{...}{Ignored.}

\item{ext}{Numeric vector c(xmin, xmax, ymin, ymax).}

\item{data}{If TRUE, return full block_data(). If FALSE (default), return wk::rct.}
}
\value{
A \code{wk::rct} vector or data.frame of intersecting blocks.
}
\description{
Find blocks intersecting an extent
}
\examples{
\dontrun{
b <- block("/path/to/raster.tif")
blocks_in_ext(b, c(100, 500, 200, 600))
}
}
